requirements.txt
# === Core ===
python-dotenv
tqdm
requests
argparse
datasets

# === LLM / Agent ===
openai
anthropic
google-generativeai
google-genai

# === Static & Dynamic Analysis Helpers ===
tree_sitter
pycparser
clang
fuzzywuzzy[speedup]

# === Validation / Execution ===
subprocess-run
pexpect

# === Data & Logging ===
pandas
numpy
pyyaml
matplotlib

# === Required packages ===
coverage
pytest
sumy
nltk
